fix(RSS-ECOMM-2_20): build fix
